+++ src/ast/passes/codegen_llvm.h
@@ -197,6 +197,7 @@ private:
                                const std::string &temp_name);
 
   void createIncDec(Unop &unop);
+  Function *createMapDeleteFilteredCallback(location loc);
 
   // Return a lambda that has captured-by-value CodegenLLVM's async id state
   // (ie `printf_id_`, `mapped_printf_id_`, etc.).  Running the returned lambda
@@ -235,6 +236,7 @@ private:
   std::map<std::string, int> next_probe_index_;
   // Used if there are duplicate USDT entries
   int current_usdt_location_index_{ 0 };
+  Function *map_delete_filtered_cb_ = nullptr;
 
   std::map<std::string, AllocaInst *> variables_;
   int printf_id_ = 0;