+++ src/types.h
@@ -49,7 +49,8 @@ enum class Type
   timestamp,
   mac_address,
   cgroup_path,
-  strerror
+  strerror,
+  map_wildcard
   // clang-format on
 };
 
@@ -372,6 +373,10 @@ public:
   {
     return type == Type::strerror;
   };
+  bool IsMapWildcardTy(void) const
+  {
+    return type == Type::map_wildcard;
+  }
 
   friend std::ostream &operator<<(std::ostream &, const SizedType &);
   friend std::ostream &operator<<(std::ostream &, Type);
@@ -432,6 +437,7 @@ SizedType CreateTimestamp();
 SizedType CreateMacAddress();
 SizedType CreateCgroupPath();
 SizedType CreateStrerror();
+SizedType CreateMapWildcard();
 
 std::ostream &operator<<(std::ostream &os, const SizedType &type);